Canny_mod is the final source for the modified Canny Edge Detector.

+++ canny_video/openCV_HW_filter.h
@@ -24,7 +24,7 @@ void Canny_HW_ARM_only( vdma_handle *vdma_handle, sepimgfilter_handle *filter_ha
 
 void Canny_HW_ARM( vdma_handle *vdma_handle, sepimgfilter_handle *filter_handle, unsigned int width, unsigned int height, unsigned int sobelIn_fbAddr, unsigned int sobelDxOut_fbAddr, unsigned int sobelDyOut_fbAddr );
 
-void Canny_HW_RC( void * destImage_virtAddr, void * dx_virtAddr, void * dy_virtAddr, unsigned int width, unsigned int height, double low_thresh, double high_thresh, int aperture_size, bool L2gradient );
+void Canny_HW_RC( void * destImage_virtAddr, void * dx_virtAddr, void * dy_virtAddr, unsigned int width, unsigned int height, int low_thresh, int high_thresh, int aperture_size, bool L2gradient );
 
 void nonMaxSuppress(int rows, int cols, int cn, Mat dx, Mat dy, ptrdiff_t mapstep, int* mag_buf[], uchar* map, int* maxsize, std::vector<uchar*> *stack, uchar*** stack_top, uchar*** stack_bottom, double threshold1, double threshold2, bool L2gradient = false );
